// src/generated/prisma/models/Order.ts
export type OrderModel = runtime.Types.Result.DefaultSelection<Prisma.$OrderPayload>
// ...
export type $OrderPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Order"
  objects: {
    user: Prisma.$UserPayload<ExtArgs>
    items: Prisma.$OrderItemPayload<ExtArgs>[]
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    user_id: string
    status: string
    payment_status: string
    total: number
    created_at: Date
    updated_at: Date
  }, ExtArgs["result"]["order"]>
  composites: {}
}