// src/generated/prisma/models/OrderItem.ts
export type OrderItemModel = runtime.Types.Result.DefaultSelection<Prisma.$OrderItemPayload>
// ...
export type $OrderItemPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "OrderItem"
  objects: {
    order: Prisma.$OrderPayload<ExtArgs>
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    order_id: string
    product_id: string | null
    name: string
    price: number
    qty: number
    variation: string | null
    image_url: string | null
  }, ExtArgs["result"]["orderItem"]>
  composites: {}
}